profil fotoğrafları webden değişebilir

+++ app/Http/Controllers/EventController.php
@@ -74,7 +74,9 @@ class EventController extends Controller
                     $event->event_content = $request->event_content;
                     if ($request->hasFile('event_image')) {
                         $event = Event::getEventInId($request->event_id);
-                        Storage::disk('public')->delete($event->event_image);
+                        if ($event->event_image){
+                            Storage::disk('public')->delete($event->event_image);
+                        }
                         $image = $request->file('event_image');
                         $filename = Str::random(40) . '.' . $image->getClientOriginalExtension();
                         $path = $image->storeAs('event_images', $filename, 'public');
@@ -98,7 +100,9 @@ class EventController extends Controller
         if (session()->has('login_control')) {
             if (session('login_control') == 1) { // daha önce login girişi yapıldı mı kontrolü yapar
                 $event = Event::getEventInId($eventId);
-                Storage::disk('public')->delete($event->event_image);
+                if ($event->event_image){
+                    Storage::disk('public')->delete($event->event_image);
+                }
                 Event::deleteEvent($eventId);
                 return redirect()->route('get-our-event-page');
             } else {
